requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Data Validation
pydantic==2.5.0

# AWS Services
boto3==1.34.0
botocore==1.34.0

# HTTP Requests
requests==2.31.0

# Audio Processing
librosa==0.10.1
numpy==1.24.3
scipy==1.11.4

# File Downloads
gdown==4.7.1

# System utilities
python-multipart==0.0.6
python-dotenv==1.0.0